Fix _parse_json_body (#33)
This pull request includes a small change to the `_parse_json_body`
function in the `src/fabric_cicd/_common/_fabric_endpoint.py` file. The
change modifies the condition to check if the response content type
contains "application/json" instead of being exactly equal to it.

*
[`src/fabric_cicd/_common/_fabric_endpoint.py`](diffhunk://#diff-df48b77e163b4a13c97f80408ad6640fca48ffdc60c19e085b4c73589933178cL240-R240):
Modified the condition in `_parse_json_body` to check if
"application/json" is in the content type header of the response.

---------

Co-authored-by: Jacob Knightley <Jacob.Knightley@microsoft.com>