Update Readme Fixes #43 (#44)
This pull request includes several documentation updates and
improvements to the `fabric-cicd` project. The most important changes
include updating the section overview in the `README.md`, renaming the
`help.md` file to `about.md`, adding installation instructions for the
private preview, and updating the navigation in the `mkdocs.yml` file.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L20-R25):
Updated the section overview links to reflect the new structure and
consolidated support and security policies under the "About" section.
* `docs/help.md` renamed to `docs/about.md`: Renamed the file to better
reflect its content, which includes support and security information.
*
[`docs/index.md`](diffhunk://#diff-b4d68dc855d0f9476d3f2ee343853bd21bf82ea9960d0cf06661baa244439dd6R25-R48):
Added installation instructions for the private preview and included a
basic example of how to use the `fabric-cicd` library.

Configuration updates:

*
[`mkdocs.yml`](diffhunk://#diff-98d0f806abc9af24e6a7c545d3d77e8f9ad57643e27211d7a7b896113e420ed2L18-R18):
Updated the navigation to reflect the renaming of the `help.md` file to
`about.md`.