Onboard UV (#48)
This pull request introduces several significant changes across various
files to improve the project setup, documentation, and configuration.
The most important changes include modifications to the GitHub Actions
workflows, updates to the project configuration files, and enhancements
to the documentation.

### Workflow and Configuration Updates:

*
[`.github/workflows/gh-pages.yml`](diffhunk://#diff-c04119feaeffc2c216f069aec66797f2f839a2074287060698e3a2440dec8d45L4-L6):
Removed the push trigger for the `main` branch to streamline the
workflow.
*
[`.github/workflows/lint.yml`](diffhunk://#diff-107e910e9f2ebfb9a741fa10b2aa7100cc1fc4f5f3aca2dfe78b905cbd73c0d2L24-R24):
Adjusted the `ruff` commands to ensure proper formatting and linting
without exiting prematurely.
[[1]](diffhunk://#diff-107e910e9f2ebfb9a741fa10b2aa7100cc1fc4f5f3aca2dfe78b905cbd73c0d2L24-R24)
[[2]](diffhunk://#diff-107e910e9f2ebfb9a741fa10b2aa7100cc1fc4f5f3aca2dfe78b905cbd73c0d2L43-R43)

### Project Configuration:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L1-R39):
Updated project dependencies and restructured the configuration to
include new dependency groups and build system requirements.
*
[`environment.yml`](diffhunk://#diff-9efd195f4e9bfb79ccd456a1d8370fafcc4bcb0b00ea3799222667d2ae818533L1-L7):
Removed the file as part of the reconfiguration process.
*
[`ruff.toml`](diffhunk://#diff-26f436cb29eecc7bb4f9066747b43df8da81ac808955ce95c109f5d7aa778989R1-R40):
Added a new configuration file for `ruff` to manage linting rules and
formatting settings.

### Documentation Enhancements:

*
[`docs/contribution.md`](diffhunk://#diff-362883943baf53434379354560bf3d11b49ca4a8eeadd1bc8e37e44b04e230e1L30-R78):
Updated the contribution guidelines to include new steps for setting up
the development environment and managing dependencies using `uv`.

### Other Changes:

*
[`.python-version`](diffhunk://#diff-652bad7c1002badcf13f8de3a5096c2724ce54ae36cd68a21ad280547dff8e87R1):
Specified Python version 3.12 for the project.
*
[`activate.ps1`](diffhunk://#diff-ae3d5b16dbd3f471d5dc8e410e67f914643e73c93bfa1fd84792be63dbde63d6R1-R45):
Added a PowerShell script to automate the setup and activation of the
virtual environment.
* `requirements-dev.txt` and `requirements-doc.txt`: Removed these files
as dependencies are now managed through `pyproject.toml`.
[[1]](diffhunk://#diff-2b4945591edfeaa4cf4d3f155e66d4b43d1bda7a55d881d5cf3107f1b05abbbcL1-L4)
[[2]](diffhunk://#diff-27d1320c75dc2fa9448e643dc192b0a611e867e88da49221acf4f754a8d63957L1-L6)

---------

Co-authored-by: Jacob Knightley <Jacob.Knightley@microsoft.com>