Fixes #64, Fixes #65 - Documentation Updates (#68)
This pull request includes a minor change to the
`src/fabric_cicd/fabric_workspace.py` file. The change updates the
description of the `repository_directory` parameter in the `__init__`
method to clarify that it refers to the local directory path of the
repository where items are to be deployed from.

*
[`src/fabric_cicd/fabric_workspace.py`](diffhunk://#diff-3bd0f70ed06c7fc7a0f77378aecc6fb108eeb5263161db3b22270f3de4df99c0L45-R45):
Updated the description of the `repository_directory` parameter in the
`__init__` method to clarify its purpose.