Onboard Unit Test Action (#99)
This pull request includes the addition of a GitHub Actions workflow for
running unit tests and an update to the `README.md` file to include a
badge for the new workflow.

### Addition of GitHub Actions workflow:

*
[`.github/workflows/unit-test.yml`](diffhunk://#diff-d4223b44df678cece85f2542c19b9cd060d43b22f753c25214d5942ef2ca5d18R1-R23):
Added a new workflow named "Tests" that runs on pull requests. The
workflow sets up Python, installs dependencies, and runs unit tests
using `pytest`.

### Update to `README.md`:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R6):
Added a badge to display the status of the new unit test workflow.