Resolves #127 - Share changelog with version check (#141)
This pull request introduces a new utility function to parse the
changelog and integrates it into the version check process. The most
important changes include the addition of the `parse_changelog`
function, updates to the `check_version` function to display changelog
entries, and the import of the `re` module.

Enhancements to changelog handling:

*
[`src/fabric_cicd/_common/_check_utils.py`](diffhunk://#diff-ed3713e24dab32c1a8de7fc474ab489877af35d51d709421fe8a708d7dc6221fR21-R89):
Added the `parse_changelog` function to parse the changelog file and
return a dictionary of versions with their changes.
*
[`src/fabric_cicd/_common/_check_utils.py`](diffhunk://#diff-ed3713e24dab32c1a8de7fc474ab489877af35d51d709421fe8a708d7dc6221fR21-R89):
Updated the `check_version` function to use `parse_changelog` and
display changelog entries for versions between the current and latest
version.

Codebase improvements:

*
[`src/fabric_cicd/_common/_check_utils.py`](diffhunk://#diff-ed3713e24dab32c1a8de7fc474ab489877af35d51d709421fe8a708d7dc6221fR7):
Imported the `re` module to support regular expression operations in the
`parse_changelog` function.