Update Mirrored Database parametization (#154)
This pull request includes an update to the documentation for the
`Mirrored Database` section. The change clarifies the behavior of the
`find_replace` section in the `parameter.yml` file.

Documentation update:

*
[`docs/how_to/item_types.md`](diffhunk://#diff-2b5373cf38b50a290bee3f403778f7dff7cdc942ddbb39026b291da113dc6025L29-R29):
Clarified that connections will always point to the original source
database unless parameterized in the `find_replace` section of the
`parameter.yml` file.