Resolves #164 Workspace ID replacement in data pipelines bug (#166)
This pull request includes a minor but important change to the
`_replace_activity_workspace_ids` method in the `fabric_workspace.py`
file. The change ensures that the type value is a string before checking
if it is found in the activities mapping.

*
[`src/fabric_cicd/fabric_workspace.py`](diffhunk://#diff-3bd0f70ed06c7fc7a0f77378aecc6fb108eeb5263161db3b22270f3de4df99c0L287-R288):
Added a type check to ensure the `activity_value` is a string before
checking if it is in the `activities_mapping`.

---------

Co-authored-by: Shira Sassoon <shirasassoon@microsoft.com>