Fixes #186 Workspace Id not updated in data pipelines (#188)
This pull request includes an enhancement to the
`src/fabric_cicd/fabric_workspace.py` file to improve the efficiency of
replacing workspace IDs in a given file. The most important change is
the modification of the `_replace_workspace_ids` method to use `re.sub`
for more efficient and concise replacement of default workspace IDs.

Codebase simplification:

*
[`src/fabric_cicd/fabric_workspace.py`](diffhunk://#diff-3bd0f70ed06c7fc7a0f77378aecc6fb108eeb5263161db3b22270f3de4df99c0L302-R311):
Modified the `_replace_workspace_ids` method to use `re.sub` for
replacing all instances of the default feature branch workspace ID with
the target workspace ID. This change simplifies the code and ensures all
matches are replaced in one operation.

Co-authored-by: Shira Sassoon <shirasassoon@microsoft.com>