Resolves #212 - Expand check_version support (#213)
This pull request includes several changes related to the management and
location of the changelog file, as well as minor updates to the
`pyproject.toml` file. The most important changes include moving the
changelog file to a new location and updating the code to reflect this
change.

Changelog management:

*
[`docs/changelog.md`](diffhunk://#diff-77f023b99d3d58008351d3e82fc06e6d06ba1bc2da9e41be6329b7fa4f419f05L3-R5):
Removed the changelog content and replaced it with an include directive
to reference the new location.
*
[`src/fabric_cicd/changelog.md`](diffhunk://#diff-375b3b0713fcf6f8212d7f9bd6eccf4ca9b73db611c469827967823acca6279aR1-R136):
Added the changelog content to this new file.
*
[`src/fabric_cicd/_common/_check_utils.py`](diffhunk://#diff-ed3713e24dab32c1a8de7fc474ab489877af35d51d709421fe8a708d7dc6221fL31-R31):
Updated the `parse_changelog` function to use the new path for the
changelog file.

Dependency updates:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L35-L48):
Added `package-data` entry to include the new changelog file.