Fix Authentication Sample (#257)
This pull request removes outdated and unconfirmed examples from the
`authentication.md` documentation file to streamline and clarify the
content. The most significant change involves the removal of release
pipeline script examples for Azure DevOps and GitHub.

### Documentation cleanup:

*
[`docs/example/authentication.md`](diffhunk://#diff-209795f61181441c3bebe319157e9b2dec863d464945eb609858eec490c73a9eL94-L134):
Removed the "Release Pipeline Script" section, which included an Azure
DevOps YAML example and a placeholder for a GitHub example. This change
simplifies the documentation by eliminating unconfirmed and redundant
examples.