Resolve CodeQL Violations (#286)
This pull request includes updates to ensure compliance with CodeQL
recommendations by adding comments explaining the use of a default
authentication mechanism when no token is provided. The changes affect
two methods in the `fabric_cicd` package.

### CodeQL compliance updates:

*
[`src/fabric_cicd/_parameter/_utils.py`](diffhunk://#diff-08e7f8767bb7d280f82e88a77521554115ad781c1dceb4ff0a7b450d38a684abL79-R79):
Added a comment in the `validate_parameter_file` method to clarify the
use of `DefaultAzureCredential` as the default authentication mechanism
when no token is provided. This ensures the public library adheres to
best practices for authentication.
*
[`src/fabric_cicd/fabric_workspace.py`](diffhunk://#diff-3bd0f70ed06c7fc7a0f77378aecc6fb108eeb5263161db3b22270f3de4df99c0L106-R106):
Added a similar comment in the `__init__` method to explain the use of
`DefaultAzureCredential` for default authentication in the
`FabricEndpoint` initialization.