v0.1.18 (#297)
This pull request updates the `fabric-cicd` package to version `0.1.18`
and includes a bug fix along with the corresponding documentation
updates.

### Version update:

*
[`src/fabric_cicd/constants.py`](diffhunk://#diff-91c2d32be351f155f090c4bd57ebe3c10e646edbc1b663e2a72b697f956a949cL7-R7):
Updated the `VERSION` constant from `0.1.17` to `0.1.18`.

### Bug fix and release notes:

*
[`src/fabric_cicd/changelog.md`](diffhunk://#diff-375b3b0713fcf6f8212d7f9bd6eccf4ca9b73db611c469827967823acca6279aR9-R14):
Added release notes for version `0.1.18`, including a fix for a bug
related to checking the environment publish state when
`publish_item_name_exclude_regex` is `None`.

---------

Co-authored-by: Shira Sassoon <shirasassoon@microsoft.com>