v0.1.21 (#350)
This pull request updates the `fabric-cicd` package to version 0.1.21,
introducing a bug fix and an internal optimization. The most important
changes include updating the version constant and documenting the new
release in the changelog.

### Version update:

*
[`src/fabric_cicd/constants.py`](diffhunk://#diff-91c2d32be351f155f090c4bd57ebe3c10e646edbc1b663e2a72b697f956a949cL7-R7):
Updated the `VERSION` constant from `"0.1.20"` to `"0.1.21"`.

### Changelog update:

*
[`src/fabric_cicd/changelog.md`](diffhunk://#diff-375b3b0713fcf6f8212d7f9bd6eccf4ca9b73db611c469827967823acca6279aR9-R15):
Added release notes for version 0.1.21, including a bug fix for
workspace ID replacement in JSON files during pipeline deployments and
an increase in the maximum retry limit for Warehouses and Dataflows.