Resolves #402 Publishing Warehouse Returns "The feature is not available" (#435)
This pull request introduces a new validation mechanism for publishing
items in Fabric CI/CD workflows, ensuring that a workspace has assigned
capacity unless the item type is exempt. It also includes updates to
constants and imports to support this functionality.

### Validation for assigned capacity:

*
[`src/fabric_cicd/constants.py`](diffhunk://#diff-91c2d32be351f155f090c4bd57ebe3c10e646edbc1b663e2a72b697f956a949cR41-R43):
Added `NO_ASSIGNED_CAPACITY_REQUIRED` constant, listing item types
(`SemanticModel`, `Report`) that do not require assigned capacity.
*
[`src/fabric_cicd/publish.py`](diffhunk://#diff-a8555a82747092e8d1f2ee36f31c947b4e5a264ed6c6d8de48d8daf850372704R55-R67):
Implemented a check in `publish_all_items` to validate if the workspace
has assigned capacity. If not, and the item type is not exempt, an error
is raised with a descriptive message.

### Supporting updates:

*
[`src/fabric_cicd/publish.py`](diffhunk://#diff-a8555a82747092e8d1f2ee36f31c947b4e5a264ed6c6d8de48d8daf850372704R9-R14):
Imported `dpath.util` for JSON path extraction and
`FailedPublishedItemStatusError` for error handling.