Resolves #447 check for items not requiring assinged capacity (#451)
This pull request updates the logic for determining whether an item
requires assigned capacity by modifying the
`NO_ASSIGNED_CAPACITY_REQUIRED` constant and updating its usage in the
publishing logic.
https://github.com/microsoft/fabric-cicd/issues/447
### Changes to constants:

*
[`src/fabric_cicd/constants.py`](diffhunk://#diff-91c2d32be351f155f090c4bd57ebe3c10e646edbc1b663e2a72b697f956a949cL42-R42):
Updated `NO_ASSIGNED_CAPACITY_REQUIRED` to be a list of lists, instead
of a flat list, to allow for more granular checks.

### Changes to publishing logic:

*
[`src/fabric_cicd/publish.py`](diffhunk://#diff-a8555a82747092e8d1f2ee36f31c947b4e5a264ed6c6d8de48d8daf850372704L64-R67):
Updated the condition in `publish_all_items` to handle the new structure
of `NO_ASSIGNED_CAPACITY_REQUIRED` by checking if
`fabric_workspace_obj.item_type_in_scope` is not in the updated list of
lists.