v0.1.32 (#667)
This pull request updates the package version to 0.1.32 and adds release
notes for this new version. The main change in this release is a bug fix
related to publishing Environment items with only spark settings.

Version update and release notes:

* Updated the `VERSION` constant in `constants.py` from "0.1.31" to
"0.1.32" to reflect the new release.
* Added release notes for version 0.1.32 in `changelog.md`, highlighting
a bug fix for publishing Environment items that contain only spark
settings.

Co-authored-by: Shira Sassoon <shirasassoon@microsoft.com>